pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "abaqusgpt"
version = "0.1.0"
description = "Your AI copilot for Abaqus: modeling, meshing, convergence debugging, and beyond."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "ZhouChaunge" }
]
keywords = [
    "abaqus",
    "fea",
    "finite-element",
    "simulation",
    "convergence",
    "ai",
    "llm",
    "agent"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
rag = [
    "llama-index>=0.10.0",
    "chromadb>=0.4.0",
]

[project.scripts]
abaqusgpt = "abaqusgpt.cli:app"

[project.urls]
Homepage = "https://github.com/ZhouChaunge/AbaqusGPT"
Repository = "https://github.com/ZhouChaunge/AbaqusGPT"
Issues = "https://github.com/ZhouChaunge/AbaqusGPT/issues"

[tool.hatch.build.targets.wheel]
packages = ["abaqusgpt"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=abaqusgpt"
